// frontend/src/lib/k8s/api/v2/ApiResource.tsx
/*
 * Copyright 2025 The Kubernetes Authors
 *
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 *
 * http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 */

/**
 * Represents a resource available in the Kubernetes API.
 */
export interface ApiResource {
  apiVersion: string;
  groupName?: string;
  version: string;
  pluralName: string;
  singularName: string;
  kind: string;
  isNamespaced: boolean;
}

/** Unique identifier for an API Resource */
export const apiResourceId = (resource: ApiResource) => JSON.stringify(resource);

/** Deserialize ApiResource ID */
export const parseFromId = (id: string): ApiResource => JSON.parse(id);